% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QProMS_dataset.R
\docType{data}
\name{QProMS_dataset}
\alias{QProMS_dataset}
\title{ProteinGroups examples.}
\format{
A data frame ..:
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}
}
\usage{
QProMS_dataset
}
\description{
Contains bla bla bla
}
\keyword{datasets}
