% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_qproms.R
\name{run_qproms}
\alias{run_qproms}
\title{Run the QProMS Shiny application}
\usage{
run_qproms(launch.browser = TRUE, port = NULL, host = "127.0.0.1")
}
\arguments{
\item{launch.browser}{If true, launch browser automatically}

\item{port}{TCP port for the server}

\item{host}{IPv4 address to bind to}
}
\description{
Launches QProMS with automatic dependency installation.
}
